+++ tests/openapi/test_tags.py
@@ -5,7 +5,7 @@ import pytest
 from starlite import Controller, HTTPRouteHandler, Router, Starlite, get
 
 if TYPE_CHECKING:
-    from openapi_schema_pydantic.v3.v3_1_0.open_api import OpenAPI
+    from pydantic_openapi_schema.v3_1_0.open_api import OpenAPI
 
 
 @pytest.fixture()